CRUD

+++ board/src/main/java/yj/springboard/board/service/BoardService.java
@@ -1,13 +1,17 @@
 package yj.springboard.board.service;
 
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import yj.springboard.board.controller.BoardController;
 import yj.springboard.board.entity.BoardEntity;
 import yj.springboard.board.repository.BoardRepository;
+
+
 @Service
 public class BoardService {
     
@@ -21,6 +25,8 @@ public class BoardService {
 
     //작성버튼을 => 데이터가 DB에 저장, 수정
     public void boardWrite(BoardEntity boardEntity){
+        boardEntity.setRegDate(new Date());
+        boardEntity.setUpdDate(boardEntity.getRegDate());
         boardRepository.save(boardEntity);
     }
 
@@ -29,8 +35,11 @@ public class BoardService {
         return boardRepository.findById(id).get();
     }
     
-    public void boardDelete(Integer id)
-    {
-        boardRepository.deleteById(id);
+    public void boardDelete(Integer id) {
+        if (boardRepository.existsById(id)) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("해당 ID의 게시글이 존재하지 않습니다.");
+        }
     }
 }
